.history/requirements_20241212144538.txt
PyMuPDF==1.19.6
tkinter==8.6
Pillow==9.5.0
mysql-connector-python==9.1.0